Modded the ili948x driver to likely compile.
I'll debug it tomorrow.

A big issue out of the gate was the use of 18bit color, which is going to
be slow on SPI.  Changed it to 16-bit color I think, although maybe more HW
config changes are needed.

Also I changed types over to int16 so that there is less needed conversion
at various points.

Very likely some bugs to sort through at this point...