Removes serial support.  Adds shell for xmodem support.
The error rate to the picocalc over its serial interface has
many errors which brings to light the fact that serial interfaces
are inheriently prone to errors.

I thnk an xmodem protocol at a minimum will be needed for
a reasonable transfer experience.  This is not implemented yet,
just the basic shell is in place.