Verified sx kind of works.
It was looking for the wrong 'c'.
I needed to use rx -c on the sender side (took a bit to figure that out)

There is a bunch of commented-out debug code for now as I want to do
more testing.

It is not truncating the last packet on the receiver side.  I'll need
to look further into how to do it correctly.